+++ packages/process/lib/Polysemy/Process/Data/ProcessError.hs
@@ -5,10 +5,15 @@ module Polysemy.Process.Data.ProcessError where
 
 import System.Exit (ExitCode)
 
+import Polysemy.Process.Data.SystemProcessError (SystemProcessScopeError)
+
 -- |Signal error for 'Polysemy.Process.Process'.
 data ProcessError =
   -- |Something broke.
   Unknown Text
+  |
+  -- |The process failed to start.
+  StartFailed SystemProcessScopeError
   |
   -- |The process terminated with exit code.
   Exit ExitCode